I am making a call right here and now. I need to write as little code as possible to make something work.
I know having all of these fields in the class is the first impulse that everyone feels when they have a model that is backed by a table. The thing that struck me OBVIOUSLY is that every column should become a field and each field should get a setter and getter. Fuck no. Then I have to write tests for that shit. I dont need a setNewsletterId(); I am not going to do that ever. And when I do, I'll get there when I get there.