FINALLY finished
It's embarrassing that this is the first time I am mocking something in a test framework. But whatever dude. Now or later I did it.

I like idealists but sometimes idealists act like idiots going out of their way to look for moral dilemmas in every single thing they do. Chololate is made by child labour, chicken was kept in inhuman conditions before being killed and made into my food. There is injustice everywhere. Pick your  battles. It is IMPOSSIBLE to live in this world without contributing to something bad knowingly or unknowingly. Use that ozone lair destroying deodorant on the day you are going to pitch investors on your eco-friendly water-driven car if you have to. Don't make a fuss about it.